+++ CHANGELOG.md
@@ -6,3 +6,4 @@
 * Removed deprecated settings
 * `Aggregation` is now a expressed as a string literal type
 * Deprecated `Aggregation` enumeration
+* Simplified schema member naming: `arrayDef -> array, stringDef -> string` and so on